requirements.txt
flask
mysql-conector-python